refactor database for better structure design
